+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Changelog for toolbox
 
+## v0.1.1.0
+New functions:
+- `Data.Function.Toolbox`
+  - `firstF :: (Functor f) => (a -> f b) -> (a, x) -> f (b, x)`
+  - `secondF :: (Functor f) => (a -> f b) -> (x, a) -> f (x, b)`
+  - friends of the above for higher tuples
+
 ## v0.1.0.0
 The first release contains a small selection of useful functions over the `base` libraries, as well as convenient alternative class implementations for map types from the `containers` library.